% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{argument}
\alias{argument}
\title{Argument}
\usage{
argument(cmd, name, nargs = 1, help = "")
}
\arguments{
\item{cmd}{Command with description, args to parse, and list of params}

\item{name}{Name of argument(s)}

\item{nargs}{Number of arguments (-1 for unlimited)}

\item{help}{Description of argument for help page}
}
\description{
Argument
}
