% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{build_help_page}
\alias{build_help_page}
\title{Help page}
\usage{
build_help_page(description, params)
}
\arguments{
\item{description}{What does the script do?}

\item{params}{List of script parameters (options and arguments)}
}
\description{
Help page
}
