% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{command}
\alias{command}
\title{Commands}
\usage{
command(description, parse = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{description}{What does the script do?}

\item{parse}{String of command line arguments}
}
\description{
Commands
}
