% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{get_arguments}
\alias{get_arguments}
\title{Get all argument params}
\usage{
get_arguments(cmd)
}
\arguments{
\item{cmd}{Command with description and list of params}
}
\description{
Get all argument params
}
