% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{get_defaults}
\alias{get_defaults}
\title{Get default values of params}
\usage{
get_defaults(cmd)
}
\arguments{
\item{cmd}{Command with description and list of params}
}
\description{
Get default values of params
}
