% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{get_nargs}
\alias{get_nargs}
\title{Get all nargs value from arguments}
\usage{
get_nargs(cmd)
}
\arguments{
\item{cmd}{Command with description and list of params}
}
\description{
Get all nargs value from arguments
}
