% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{interval}
\alias{interval}
\title{Interval}
\usage{
interval(lower, upper, exclusive = FALSE, exclude_lower = FALSE,
  exclude_upper = FALSE)
}
\arguments{
\item{lower}{Lower bound of interval}

\item{upper}{Upper bound of interval}

\item{exclusive}{Exclude bounds?}

\item{exclude_lower}{Exclude lower bound?}

\item{exclude_upper}{Exclude upper bound?}
}
\description{
Interval
}
