% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if numeric value is an integer
(Does not use the name is.integer because it is already defined in base)}
\usage{
is.wholenumber(x)
}
\arguments{
\item{x}{Numeric}
}
\description{
Check if numeric value is an integer
(Does not use the name is.integer because it is already defined in base)
}
