% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{is_valid_interval_value}
\alias{is_valid_interval_value}
\title{Check validity of interval value}
\usage{
is_valid_interval_value(param, value)
}
\arguments{
\item{param}{Script parameter}

\item{value}{Value of parameter}
}
\description{
Check validity of interval value
}
