% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{match_positionals_with_arguments}
\alias{match_positionals_with_arguments}
\title{Match the positional args given with those defined in command.
At most one argument may allow for a variable number of args (nargs = Inf),
so cycle through the args forward (until Inf is encountered), then,
if necessary, backward to fully identify the begin/end of the Inf arg.}
\usage{
match_positionals_with_arguments(args, total_nargs)
}
\arguments{
\item{args}{Arguments}

\item{total_nargs}{Total number of arguments provided by the user}
}
\description{
Match the positional args given with those defined in command.
At most one argument may allow for a variable number of args (nargs = Inf),
so cycle through the args forward (until Inf is encountered), then,
if necessary, backward to fully identify the begin/end of the Inf arg.
}
