% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_lists}
\alias{merge_lists}
\title{Helper function to merge lists (left to right)}
\usage{
merge_lists(...)
}
\arguments{
\item{...}{Lists to merge}
}
\description{
Helper function to merge lists (left to right)
}
