% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{one_getopt_param}
\alias{one_getopt_param}
\title{Translate a single parameter element to the format required for parse_args}
\usage{
one_getopt_param(param)
}
\arguments{
\item{param}{a list containing parameter details}
}
\description{
Translate a single parameter element to the format required for parse_args
}
