% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{parse_parameter_type}
\alias{parse_parameter_type}
\title{Parse parameter type}
\usage{
parse_parameter_type(param, value)
}
\arguments{
\item{param}{Script parameter}

\item{value}{Value of parameter}
}
\description{
Parse parameter type
}
