% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{A command and an argument, option, or script to apply to it}
}
\description{
scriptr is designed for use with magrittr's pipe function, \code{\%>\%},
to turn function composition into a series of imperative statements.
}
\examples{
# Instead of
script(option(command("Powers of 2"), "--pow", default = 1), function(pow) print(2^pow))
# you can write
command("Powers of 2") \%>\% option("--pow", default = 1) \%>\% script(function(pow) print(2^pow))
}
