% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{prepare_getopt_param}
\alias{prepare_getopt_param}
\title{Translate a single parameter element to the format required for getopt}
\usage{
prepare_getopt_param(param)
}
\arguments{
\item{param}{List containing parameter details}
}
\description{
Translate a single parameter element to the format required for getopt
}
