% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{process_getopt_values}
\alias{process_getopt_values}
\title{Convert strings returned by getopt into appropriate R types}
\usage{
process_getopt_values(params, values)
}
\arguments{
\item{params}{List of script parameters}

\item{values}{List of script values in strings}
}
\description{
Convert strings returned by getopt into appropriate R types
}
