% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Simple function to remove option prefix}
\usage{
remove_prefix(x)
}
\arguments{
\item{x}{String beginning with '-' or '--'}
}
\description{
Simple function to remove option prefix
}
