% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{script}
\alias{script}
\title{Script}
\usage{
script(cmd, fun)
}
\arguments{
\item{cmd}{Command with description and list of params}

\item{fun}{Function to execute with arguments supplied from the command line}
}
\description{
Script
}
