% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr-package.R
\docType{package}
\name{scriptr}
\alias{scriptr}
\alias{scriptr-package}
\title{scriptr - parse command line arguments from the shell}
\description{
scriptr - parse command line arguments from the shell
}
