% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{validate_choice_value}
\alias{validate_choice_value}
\title{Validate value of choice type}
\usage{
validate_choice_value(param, value)
}
\arguments{
\item{param}{Script parameter}

\item{value}{Value of parameter}
}
\description{
Validate value of choice type
}
