% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{validate_interval_value}
\alias{validate_interval_value}
\title{Validate value of interval type}
\usage{
validate_interval_value(param, value)
}
\arguments{
\item{param}{Script parameter}

\item{value}{Value of parameter}
}
\description{
Validate value of interval type
}
