% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{argument}
\alias{argument}
\title{Argument}
\usage{
argument(cmd, name, type = "character", nargs = 1, help = "")
}
\arguments{
\item{cmd}{Command with description and list of params}

\item{name}{Name of argument}

\item{type}{String of data type, scriptr::interval, or scriptr::choice}

\item{nargs}{Number of arguments (Inf for unlimited)}

\item{help}{Description of argument for help page}
}
\description{
Argument
}
