% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{option}
\alias{option}
\title{Option}
\usage{
option(cmd, ..., default = NULL, type = NULL, is.flag = FALSE,
  help = "")
}
\arguments{
\item{cmd}{Command with description and list of params}

\item{...}{Long and short options}

\item{default}{Default option value if none is given}

\item{type}{String of atomic data type, scriptr::interval, or scriptr::choice}

\item{is.flag}{Is this a simple logical flag?}

\item{help}{Description of option for help page}
}
\description{
Option
}
