% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{parse_args}
\alias{parse_args}
\title{Parse command-line arguments using standard libc functions}
\usage{
parse_args(scp, args)
}
\arguments{
\item{scp}{Script object}

\item{args}{Command line args}
}
\description{
Parse command-line arguments using standard libc functions
}
