% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{parse_args}
\alias{parse_args}
\title{Parse command-line arguments using standard libc functions}
\usage{
parse_args(params, args)
}
\arguments{
\item{cmd}{command definition, of the form list(a='long-a',b='long-b')}

\item{argv}{arguments to parse (excluding command name)}
}
\description{
Parse command-line arguments using standard libc functions
}
