% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getopt.R
\name{parse_args}
\alias{parse_args}
\title{Parse command-line arguments using standard libc functions}
\usage{
parse_args(params, args)
}
\arguments{
\item{params}{List of script parameters}

\item{args}{Command line arguments}
}
\description{
Parse command-line arguments using standard libc functions
}
