% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{A command and an argument, option, or script to apply to it}
}
\description{
scriptr is designed for use with magrittr's pipe function, \code{\%>\%},
to turn function composition into a series of imperative statements.
}
